requirements.txt
pandas==2.2.2
openpyxl==3.1.5
jinja2==3.1.4
python-dotenv==1.0.1
email-validator==2.2.0
rich==13.7.1
